- Revert: Change index parameter types from int to size_t to make 0-based range implicit. - Add checks that index values are >= 0.
git-svn-id: https://chromiumembedded.googlecode.com/svn/trunk@409 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ include/cef.h
@@ -2663,8 +2663,8 @@ public:
   ///
   // Returns true if the object has a value with the specified identifier.
   ///
-  /*--cef(capi_name=has_value_byindex)--*/
-  virtual bool HasValue(size_t index) =0;
+  /*--cef(capi_name=has_value_byindex,index_param=index)--*/
+  virtual bool HasValue(int index) =0;
 
   ///
   // Delete the value with the specified identifier.
@@ -2674,8 +2674,8 @@ public:
   ///
   // Delete the value with the specified identifier.
   ///
-  /*--cef(capi_name=delete_value_byindex)--*/
-  virtual bool DeleteValue(size_t index) =0;
+  /*--cef(capi_name=delete_value_byindex,index_param=index)--*/
+  virtual bool DeleteValue(int index) =0;
 
   ///
   // Returns the value with the specified identifier.
@@ -2685,8 +2685,8 @@ public:
   ///
   // Returns the value with the specified identifier.
   ///
-  /*--cef(capi_name=get_value_byindex)--*/
-  virtual CefRefPtr<CefV8Value> GetValue(size_t index) =0;
+  /*--cef(capi_name=get_value_byindex,index_param=index)--*/
+  virtual CefRefPtr<CefV8Value> GetValue(int index) =0;
 
   ///
   // Associate a value with the specified identifier.
@@ -2697,8 +2697,8 @@ public:
   ///
   // Associate a value with the specified identifier.
   ///
-  /*--cef(capi_name=set_value_byindex)--*/
-  virtual bool SetValue(size_t index, CefRefPtr<CefV8Value> value) =0;
+  /*--cef(capi_name=set_value_byindex,index_param=index)--*/
+  virtual bool SetValue(int index, CefRefPtr<CefV8Value> value) =0;
 
   ///
   // Register an identifier whose access will be forwarded to the CefV8Accessor
@@ -3137,8 +3137,8 @@ public:
   ///
   // Returns the value of the attribute at the specified 0-based index.
   ///
-  /*--cef(capi_name=get_attribute_byindex)--*/
-  virtual CefString GetAttribute(size_t index) =0;
+  /*--cef(capi_name=get_attribute_byindex,index_param=index)--*/
+  virtual CefString GetAttribute(int index) =0;
 
   ///
   // Returns the value of the attribute with the specified qualified name.
@@ -3182,8 +3182,8 @@ public:
   // Moves the cursor to the attribute at the specified 0-based index. Returns
   // true if the cursor position was set successfully.
   ///
-  /*--cef(capi_name=move_to_attribute_byindex)--*/
-  virtual bool MoveToAttribute(size_t index) =0;
+  /*--cef(capi_name=move_to_attribute_byindex,index_param=index)--*/
+  virtual bool MoveToAttribute(int index) =0;
 
   ///
   // Moves the cursor to the attribute with the specified qualified name.
